host: Add timeout when getting a job stream.
Previously, if a job did not exist, the connection would wait forever.

The new behavior is still to wait for a job to be created (support for asking for events of a job that hasn't been created yet is part of standard usage, since job event streams are unbuffered and thus attaching first is actually necessary to get the full stream)... but do not wait forever.

Signed-off-by: Eric Myhre <eric@flynn.io>